requirements.txt
google-cloud-bigquery==2.6.2
apache-airflow-providers-redis==1.0.1
requests~=2.23.0
PyYAML~=5.4.1